Anomaly_detection/requirements.txt
discord==2.3.2
twilio==9.0.2
replicate==0.25.1
openai==1.14.2
pyserial==3.5
joblib==1.3.2
scikit-learn==1.4.1.post1
